Remove configuredTestCoverage + change docs of checkedTestCoverage + improve docs in complete example
