+++ index.js
@@ -1,6 +1,5 @@
 module.exports = {
 
-  ConfiguredTestCoverage: require('./src/ConfiguredTestCoverage'),
   ExecutedLint: require('./src/ExecutedLint'),
   ExecutedTestCoverage: require('./src/ExecutedTestCoverage'),
   ExecutedTestCoverageCheck: require('./src/ExecutedTestCoverageCheck'),
