+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@cuties/wall",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "Quality checker for projects for browser js and Node.js, can be integrated in build process of Page, also can be as stand alone application.",
   "main": "index.js",
   "scripts": {
